% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_parameters.R
\name{r21_booster_profile}
\alias{r21_booster_profile}
\title{R21 booster vaccine profile}
\usage{
r21_booster_profile
}
\description{
Parameters for a booster dose of R21 for use with the
set_mass_pev and set_pev_epi functions (Schmit + Topazian et al. 2022 Lancet ID)
}
