% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_parameters.R
\name{rtss_booster_profile}
\alias{rtss_booster_profile}
\title{RTS,S booster vaccine profile}
\usage{
rtss_booster_profile
}
\description{
Parameters for a booster dose of RTS,S for use with the
set_mass_pev and set_pev_epi functions (White MT et al. 2015 Lancet ID)
}
